% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getApiFootball.R
\name{API-Football}
\alias{API-Football}
\alias{getApiFootball}
\alias{addApiFootball}
\alias{updateApiFootball}
\title{Get data from ApiFootball.}
\usage{
getApiFootball(sport, datatype, api.key = NULL, ...)

addApiFootball(sport, datatype, api.key, all = FALSE, ...)

updateApiFootball(update.games = TRUE, update.teams = FALSE, ...)
}
\arguments{
\item{sport}{\code{Character(1L)}. Sport to be fetched.}

\item{datatype}{\code{Character(1L)}. Datatype to be fetched.}

\item{api.key}{\code{Character(1L)}. Api key to Api-Football.}

\item{all}{\code{Logical(1L)}. Whether to get leagues and seasons from
database and fetch results on all of them. If \code{FALSE}, you have to
specify league and season, for instance, when fetching games.}
}
\value{
Fetched data or path to database file.
}
\description{
Get data from ApiFootball.
}
\details{
With this function, you can fetch data from
[API-Football](https://www.api-football.com/sports#).

\code{getApiFootball} returns the results while \code{addApiFootball} adds
them to SQL database. \code{updateApiFootball} can bet utilized to
to fetch all available data. It is especially useful if you want to update
your local database daily; just run \code{updateFootballApi} and it
updates the database for you.

By default, the database is stored to directory
\code{"data/apifootball")}. You can cpntrpl this with
\code{dir.name}.
}
\examples{

# Get games
res <- getApiFootball("basketball", "games", league = 1, season = 2023, api.key = "7183699be7198ae838421831d788e077")

# Fetch leagues and add them to database
addApiFootball("basketball", "leagues", api.key = "7183699be7198ae838421831d788e077")

# Update the database
updateApiFootball(api.key = "7183699be7198ae838421831d788e077")

}
